move ensimeCompileOnly into ensime extras
