+++ src/sbt-test/sbt-ensime/compile-only/build.sbt
@@ -1,4 +1,4 @@
 val a = project
 val b = project.settings(
-  scalacOptions in (Test, EnsimeKeys.ensimeCompileOnly) ++= Seq("-Xshow-phases")
+  scalacOptions in (Test, ensimeCompileOnly) ++= Seq("-Xshow-phases")
 )
